Revise page structure to include API changes
