+++ app/(old)/hotel/[hotelID]/page.tsx
