+++ app/hotels/[hotelId]/[roomId]/page.tsx
