+++ app/hotels/[hotelId]/page.tsx
